refactor(lotto): 로또 같은 숫자 비교 결과 자료형 int -> long 변경
